saner

+++ scripts/runner/setup_aarch64-Linux.sh
@@ -133,17 +133,19 @@ else
    return 1 || exit 1
   fi
  ##Check for Minisign
-  if [ -n "${MINISIGN_KEY+x}" ] && [ -n "${MINISIGN_KEY##*[[:space:]]}" ]; then
-    mkdir -p "${HOME}/.minisign" && \
-    echo "pkgforge-minisign: minisign encrypted secret key" > "${HOME}/.minisign/pkgforge.key" &&\
-    echo "${MINISIGN_KEY}" >> "${HOME}/.minisign/pkgforge.key"
-   #https://github.com/pkgforge/.github/blob/main/keys/minisign.pub
-    export MINISIGN_PUB_KEY='RWSWp/oBUfND5B2fSmDlYaBXPimGV+r2s9skVRYTQ5cJ+7i6ff/1Nxcr'
-  else
-    echo -e "\n[-] MINISIGN_KEY is NOT Exported"
-    echo -e "Export it to Use minisign (Signing)\n"
-   export CONTINUE="NO"
-   return 1 || exit 1
+  if [[ ! -s "${HOME}/.minisign/pkgforge.key" || $(stat -c%s "${HOME}/.minisign/pkgforge.key") -le 10 ]]; then
+    if [ -n "${MINISIGN_KEY+x}" ] && [ -n "${MINISIGN_KEY##*[[:space:]]}" ]; then
+      mkdir -p "${HOME}/.minisign" && \
+      echo 'pkgforge-minisign: minisign encrypted secret key' > "${HOME}/.minisign/pkgforge.key" &&\
+      echo "${MINISIGN_KEY}" >> "${HOME}/.minisign/pkgforge.key"
+     #https://github.com/pkgforge/.github/blob/main/keys/minisign.pub
+      export MINISIGN_PUB_KEY='RWSWp/oBUfND5B2fSmDlYaBXPimGV+r2s9skVRYTQ5cJ+7i6ff/1Nxcr'
+    else
+      echo -e "\n[-] MINISIGN_KEY is NOT Exported"
+      echo -e "Export it to Use minisign (Signing)\n"
+     export CONTINUE="NO"
+     return 1 || exit 1
+    fi
   fi
 fi
 #-------------------------------------------------------#
@@ -252,7 +254,7 @@ if [ "${CONTINUE}" == "YES" ]; then
   #Test
   if ! command -v crystal &> /dev/null; then
    echo -e "\n[-] crystal NOT Found\n"
-   export CONTINUE="NO"
+   #export CONTINUE="NO"
    #return 1 || exit 1
   else
    crystal --version ; shards --version
